requirements.txt
Django==4.1.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pymongo==4.7.1
drf-spectacular==0.27.2
